backend/requirements.txt
Flask==3.1.1
flask-pymongo==3.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
Flask-Cors>=3.0.0
logging==0.4.9.6